Add files via upload
This code aims to provide patent-based information on technologies and patenting trends in football boots/cleat/shoes. It covers patent applications, granted patents, countries of origin and priority revealing interesting facets of related research and innovation. The concept of football boots has been there for a very long time. 
 However, over the years, they have kept evolving like every other part of the sports due to better research and advancement in technology. The data was sourced from the Lens data base using patent searching techniques to get patents related to football boots/cleat/shoes.

Title word count:
The title of a patent can reveal the technical areas in which patent applications are being filed. A count of recurring words in the title column shed light in the areas in which patents are being filed.The count will look at the top 100 words with more than 3 letters.

Sample Size:
The  total number of applications was determined to get the sample size

Patent Documents:
 Types of patents documents in samples was determined.
 This will shed light on the break down on granted patents, patent applications etc

Legal Status:
 Determining the legal status vs count of patent documents vs publication year.
 This will show which patents are active, expired, Pending etc

Priority Year:
Split the year from Earliest_Priority_Date column to get the priority year. Innovation is secured by a patent from the priority date. Patents are granted for period of 20 years from the priority date. If you want to know when a patent expires, you need to take the priority date and add 20 years.

Patent Jurisdiction:
Number of patent publication per country by year. This will show the countries with potential for protection based on low numbers and country which are saturated.

 Top patent applicants:
 This shows the companies most active in the field of football boots/cleat/shoes. Some of the Applicants/Companies have patented via different subsidiaries such as Nike (Subsidiaries: Nike INC, Nike Innovate etc). Applicants name harmonization is needed, harmonization is done to the most common applicants.

Top owners of patents:  
Some patents are owned by multiple owners. Harmonization of owners is needed.

The top CPC and IPCR classes:
A classification scheme is a system of codes that groups inventions according to technical area, which means similar inventions are grouped under the same classification. One invetion can be classed in more than one class. A CTE was used to split classes and to do a count.

 --Example of a Main class with its subclasses:
       -- A43  reprents the main class called 'FOOTWEAR' 
       -- A43B  is a subclass of A43 for 'CHARACTERISTIC FEATURES OF FOOTWEAR PARTS OF FOOTWEAR'
       --A43B23/00 is a subclass of A43B about 'Uppers Boot legs Stiffeners Other single parts of footwear'
       --A43B23/02  is a subclass of Uppers Boot legs

Priority year vs Patent Priority Count: 
 Patents are normally granted for a duration of 20 years. Looked at where patent priority is concentrated to determine the maturity